# [Replicated] all: host a DRPC server and use it for node-node BatchRequests