--- a/pkg/sql/test_file_226.go
+++ b/pkg/sql/test_file_226.go
@@ -1,0 +1,12 @@
+
+    // Package sql
+    package sql
+
+    // TestFunction is a sample test function created for commit b8115b4b
+    func TestFunction() {
+        // Test implementation
+        // Original commit SHA: b8115b4b92984dd257423993a195ee47103b4cb9
+        // Added on: 2025-01-17T11:05:15.706035
+        // This is a single file change for demonstration
+    }
+