--- a/pkg/sql/test_file_310.go
+++ b/pkg/sql/test_file_310.go
@@ -1,0 +1,12 @@
+
+    // Package sql
+    package sql
+
+    // TestFunction is a sample test function created for commit 6a2a8132
+    func TestFunction() {
+        // Test implementation
+        // Original commit SHA: 6a2a8132e1ffff295df5b7c0e839d6a96ee3f7bf
+        // Added on: 2025-01-17T11:04:59.147604
+        // This is a single file change for demonstration
+    }
+