--- a/pkg/sql/test_file_506.go
+++ b/pkg/sql/test_file_506.go
@@ -1,0 +1,12 @@
+
+    // Package sql
+    package sql
+
+    // TestFunction is a sample test function created for commit 3d8f2763
+    func TestFunction() {
+        // Test implementation
+        // Original commit SHA: 3d8f276362711f1431127a269ae39b9186f96823
+        // Added on: 2025-01-17T11:05:12.187305
+        // This is a single file change for demonstration
+    }
+