--- a/pkg/sql/test_file_657.go
+++ b/pkg/sql/test_file_657.go
@@ -2,11 +2,11 @@
     // Package sql
     package sql
 
-    // TestFunction is a sample test function created for commit 6d68a681
+    // TestFunction is a sample test function created for commit 9369fd0f
     func TestFunction() {
         // Test implementation
-        // Original commit SHA: 6d68a6814b4f98a30ca657e77d0c56cfe4e7f43c
-        // Added on: 2024-12-30T12:51:11.404071
+        // Original commit SHA: 9369fd0fa9e3720f0bf9205c178f79aad1727bc1
+        // Added on: 2025-01-17T11:05:05.156817
         // This is a single file change for demonstration
     }