--- a/pkg/sql/test_file_684.go
+++ b/pkg/sql/test_file_684.go
@@ -1,0 +1,12 @@
+
+    // Package sql
+    package sql
+
+    // TestFunction is a sample test function created for commit 917b53f9
+    func TestFunction() {
+        // Test implementation
+        // Original commit SHA: 917b53f9f4bcad346f212a6552e08f6c2082998e
+        // Added on: 2025-01-17T11:04:53.425971
+        // This is a single file change for demonstration
+    }
+