--- a/pkg/sql/test_file_684.go
+++ b/pkg/sql/test_file_684.go
@@ -2,11 +2,11 @@
     // Package sql
     package sql
 
-    // TestFunction is a sample test function created for commit 917b53f9
+    // TestFunction is a sample test function created for commit b0ea57cf
     func TestFunction() {
         // Test implementation
-        // Original commit SHA: 917b53f9f4bcad346f212a6552e08f6c2082998e
-        // Added on: 2025-01-17T11:04:53.425971
+        // Original commit SHA: b0ea57cf550d1cb039d34221f9613ccbfb23da28
+        // Added on: 2025-01-17T11:05:22.988133
         // This is a single file change for demonstration
     }