--- a/pkg/sql/test_file_761.go
+++ b/pkg/sql/test_file_761.go
@@ -1,0 +1,12 @@
+
+    // Package sql
+    package sql
+
+    // TestFunction is a sample test function created for commit 053101a7
+    func TestFunction() {
+        // Test implementation
+        // Original commit SHA: 053101a7f31ea4edfdd9cae2689efa34cf73466a
+        // Added on: 2025-01-17T11:04:56.265454
+        // This is a single file change for demonstration
+    }
+