--- a/pkg/sql/test_file_911.go
+++ b/pkg/sql/test_file_911.go
@@ -1,0 +1,12 @@
+
+    // Package sql
+    package sql
+
+    // TestFunction is a sample test function created for commit 42822782
+    func TestFunction() {
+        // Test implementation
+        // Original commit SHA: 428227821643c53ae8dfbb8976f21de4eaae950e
+        // Added on: 2025-01-17T11:05:19.859215
+        // This is a single file change for demonstration
+    }
+