--- a/pkg/sql/test_file_954.go
+++ b/pkg/sql/test_file_954.go
@@ -1,0 +1,12 @@
+
+    // Package sql
+    package sql
+
+    // TestFunction is a sample test function created for commit b0ae3171
+    func TestFunction() {
+        // Test implementation
+        // Original commit SHA: b0ae3171018e556096f005af3ff4930f5ac8d143
+        // Added on: 2025-01-17T11:05:09.196625
+        // This is a single file change for demonstration
+    }
+