--- a/pkg/sql/test_file_970.go
+++ b/pkg/sql/test_file_970.go
@@ -1,0 +1,12 @@
+
+    // Package sql
+    package sql
+
+    // TestFunction is a sample test function created for commit 17231906
+    func TestFunction() {
+        // Test implementation
+        // Original commit SHA: 17231906e28a0ee4e374139299941e1ec3257f7f
+        // Added on: 2025-01-17T11:05:02.019450
+        // This is a single file change for demonstration
+    }
+