Replicated from original PR #138368

Original author: cthumuluru-crdb
Original creation date: 2025-01-07T10:59:09Z

Original reviewers: tbg, cthumuluru-crdb

Original description:
---
drpc (github.com/storj/drpc) is a lightweight, drop-in replacement for gRPC.  Initial benchmark results from the Perf Efficiency team, using a simple ping service (github.com/cockroachlabs/perf-efficiency-team/tree/main/rpctoy),
demonstrate that switching from gRPC to drpc delivers significant performance improvements.

```
$ benchdiff --old lastmerge ./pkg/sql/tests -b -r 'Sysbench/SQL/3node/oltp_read_write' -d 1000x -c 10

name                                   old time/op    new time/op    delta
Sysbench/SQL/3node/oltp_read_write-24    14.0ms ± 2%    14.1ms ± 1%     ~     (p=0.063 n=10+10)

name                                   old alloc/op   new alloc/op   delta
Sysbench/SQL/3node/oltp_read_write-24    2.55MB ± 1%    2.26MB ± 2%  -11.39%  (p=0.000 n=9+9)

name                                   old allocs/op  new allocs/op  delta
Sysbench/SQL/3node/oltp_read_write-24     17.2k ± 1%     11.7k ± 0%  -32.30%  (p=0.000 n=10+8)
```

This pull request introduces experimental support for using drpc to serve BatchRequests. The feature is disabled by default but can be enabled explicitly by setting the `COCKROACH_EXPERIMENTAL_DRPC_ENABLED` environment variable or automatically in CRDB test builds. The prototype lacks key features, such as authorization checks, interceptors, etc, and is not production-ready.

When DRPC is disabled, sever, muxer and listener mock implementations are injected. If it is enabled real implementations are used. DRPC is only used for BatchRequests. This PR also added support for pooled streaming unary operations similar to https://github.com/cockroachdb/cockroach/pull/136648.

Added unit tests to ensure:
- CRDB nodes can server BatchRequests.
- Simple queries when DRPC is enabled or disabled.

Note: drpc protocol generation is not yet integrated into the build process. Protobuf files were generated manually and committed directly to the repository.

This PR has a few parts:

- it adds a copy of drpc generated code for a service that supports unary BatchRequest.
- in `rpc.NewServerEx`, we now also set up a drpc server and return it up the stack.
- Registers BatchRequest service with both DRPC and gRPC. If DRPC is enabled, it is
used to make BatchRequests, if not gRPC is used.
- in our listener setup, we configure cmux to match on the `DRPC!!!1` header and serve the resulting listener on the drpc server. The drpc example uses `drpcmigrate.ListenMux` instead of cmux; we keep cmux but must make sure the header is read off the connection before delegating (which the drpxmigrate mux would have done for us).
- if using TLS, wrap the drpc listener with TLS config and use it to servr DRPC requests.
- add support to reuse drpc streams across unary BatchRequests. However, the DRPC implementation is not on par with the gRPC counterpart in terms of allocation optimizations.

Closes https://github.com/cockroachdb/cockroach/pull/136794.
Touches #137252.

Epic: CRDB-42584
Release note: None